Add transposition of connection matrices
Transpose connection matrices for dopamine, excitator, and inhibitor. Taking the transpose of the connectivity matrix is necessary because in the model the simulated rate corresponds to the post-synaptic neuron. This means that any dopaminergic modulation applied in a given region (for example, the Substantia Nigra) should propagate toward the regions that receive input from it, namely the striatum (putamen and caudate).